requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.3.3
WTForms==3.0.1
